% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_solution.R
\name{exact_control}
\alias{exact_control}
\title{The exact optimal control for the HJB PDE problem}
\usage{
exact_control(mu, rate, volat)
}
\arguments{
\item{mu}{the mean return of the risky asset}

\item{rate}{the risk-free rate of return of the risk-free asset}

\item{volat}{the volatility of the risky asset}
}
\value{
numeric
}
\description{
{The optimal control is the so-called Kelly fraction \eqn{(\mu-r)/\sigma}. }
}
