% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_solution.R
\name{exact_sol}
\alias{exact_sol}
\title{The exact solution to HJB-PDE problem for optimal log-utility}
\usage{
exact_sol(x, t, time_length, mu, rate, volat)
}
\arguments{
\item{x}{the current wealth level}

\item{t}{the current time}

\item{time_length}{the investment time-horizon}

\item{mu}{the mean rate of return of the stock}

\item{rate}{the risk-free rate of the money-market account}

\item{volat}{the annual volatility of the stock}
}
\value{
vector containing the value, the first derivative, second derivative and initial bankroll
}
\description{
{The exact solution to the Hamilton-Jacobi-Bellman PDE for the maximal log utility value of a portfolio on a
geometric Brownian motion and risk-free money-market account.}
}
