% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{imex_control}
\alias{imex_control}
\title{The numerical optimal control value}
\usage{
imex_control(
  bankroll,
  time_length,
  mu,
  rate,
  volat,
  space_epsilon = 0.001,
  time_res = 1000L,
  space_res = 400L
)
}
\arguments{
\item{bankroll}{the initial wealth level}

\item{time_length}{the time-horizon of the investment}

\item{mu}{the mean rate of return of the stock}

\item{rate}{the risk-free rate of the money-market account}

\item{volat}{the annual volatility of the stock}

\item{space_epsilon}{the truncation towards zero of the space (wealth) variable}

\item{time_res}{the time resolution for the time-grid}

\item{space_res}{the space resolution for the space-grid}
}
\value{
matrix
}
\description{
{Numerically compute the optimal control function/value of the HJB PDE for optimal log utility.}
}
