% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_wrappers.R
\name{imex_integrator}
\alias{imex_integrator}
\title{The numerical solution to HJB-PDE problem for optimal log-utility}
\usage{
imex_integrator(
  bankroll,
  time_length,
  mu,
  rate,
  volat,
  output = "value",
  space_epsilon = 0.001,
  time_res = 1000,
  space_res = 400
)
}
\arguments{
\item{bankroll}{the initial bankroll to invest with}

\item{time_length}{the time-horizon of the investment}

\item{mu}{the mean rate of return of the stock}

\item{rate}{the risk-free rate of the money-market account}

\item{volat}{the annual volatility of the stock}

\item{output}{string to choose output: 'value' or 'grid'}

\item{space_epsilon}{the truncation towards zero of the space (wealth) variable}

\item{time_res}{the time resolution for the time-grid}

\item{space_res}{the space resolution for the space-grid}
}
\value{
vector
}
\description{
{Numerically solve the Hamilton-Jacobi-Bellman PDE for the maximal log utility value of a portfolio on a
geometric Brownian motion and risk-free money-market account. An implicit-explicit scheme is used, where the
linear part of the differential equation is handled implicitly and the non-linear term is handled explicitly in time.}
}
