% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{imex_integrator1}
\alias{imex_integrator1}
\title{The numerical solution to HJB-PDE problem for optimal log-utility}
\usage{
imex_integrator1(
  bankroll,
  time_length,
  mu,
  rate,
  volat,
  space_epsilon = 0.001,
  time_res = 1000L,
  space_res = 400L
)
}
\arguments{
\item{bankroll}{the initial bankroll to invest with}

\item{time_length}{the time-horizon of the investment}

\item{mu}{the mean rate of return of the stock}

\item{rate}{the risk-free rate of the money-market account}

\item{volat}{the annual volatility of the stock}

\item{space_epsilon}{the truncation towards zero of the space (wealth) variable}

\item{time_res}{the time resolution for the time-grid}

\item{space_res}{the space resolution for the space-grid}
}
\value{
vector
}
\description{
{Numerically solve to the Hamilton-Jacobi-Bellman PDE for the maximal log utility value of a portfolio on a
geometric Brownian motion and risk-free money-market account.}
}
