% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogUtility-package.R
\docType{package}
\name{LogUtility}
\alias{LogUtility}
\alias{LogUtility-package}
\title{LogUtility: Solve the HJB PDE for optimal log utility for portfolio of risky stock and risk-free bond}
\description{
Solve numerically and exactly the HJB PDE for optimal log utility for portfolio of risky stock 
        and risk-free bond.
}
\keyword{internal}
